# docs/api.yml
openapi: 3.0.0
paths:
...
  /users-member-count:
    get:
      operationId: getUserMemberCount
      parameters: []
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Count'
        '400':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseError'
        '403':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseError'
        '404':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseError'
        '500':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseError'
      security:
        - bearer: []
      summary: ''
      tags:
        - Users
  /users:
    get:
      operationId: getUsers
      parameters:
        - name: q
          required: false
          in: query
          schema:
            type: string
        - name: role
          required: false
          in: query
          schema:
            enum:
              - ADMIN
              - REGION_MANAGER
              - COMMITTEE_MANAGER
              - ASSOCIATION_MANAGER
              - SIMPLE_USER
            type: string
        - name: nic
          required: false
          in: query
          schema:
            type: string
        - name: apv
          required: false
          in: query
          schema:
            type: string
        - name: lastName
          required: false
          in: query
          schema:
            type: string
        - name: firstName
          required: false
          in: query
          schema:
            type: string
        - name: username
          required: false
          in: query
          schema:
            type: string
        - name: responsabilityId
          required: false
          in: query
          schema:
            type: string
        - name: sacramentId
          required: false
          in: query
          schema:
            type: string
        - name: regionId
          required: false
          in: query
          schema:
            type: string
        - name: committeeId
          required: false
          in: query
          schema:
            type: string
        - name: associationId
          required: false
          in: query
          schema:
            type: string
        - name: gender
          required: false
          in: query
          schema:
            enum:
              - MALE
              - FEMALE
            type: string
        - name: page
          required: false
          in: query
          schema:
            type: number
            minimum: 1
            default: 1
        - name: pageSize
          required: false
          in: query
          schema:
            type: number
            default: 10
            minimum: 1
            maximum: 500
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
        '400':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseError'
        '403':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseError'
        '404':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseError'
        '500':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseError'
      security:
        - bearer: []
      summary: ''
      tags:
        - Users
    post:
      operationId: createUser
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUser'
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseError'
        '403':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseError'
        '404':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseError'
        '500':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseError'
      security:
        - bearer: []
      summary: ''
      tags:
        - Users
  /users/{id}:
    get:
      operationId: getUserById
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseError'
        '403':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseError'
        '404':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseError'
        '500':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseError'
      security:
        - bearer: []
      summary: ''
      tags:
        - Users
    delete:
      operationId: deleteUserById
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseError'
        '403':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseError'
        '404':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseError'
        '500':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseError'
      security:
        - bearer: []
      summary: ''
      tags:
        - Users
  /users/{id}/picture/raw:
    put:
      operationId: updateProfilePicture
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        description: Upload a user profile picture
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UploadeSuccessResponse'
        '400':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseError'
        '403':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseError'
        '404':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseError'
        '500':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseError'
      security:
        - bearer: []
      summary: ''
      tags:
        - Users
  /users/{id}/infos:
    put:
      operationId: updateUserInfo
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUser'
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseError'
        '403':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseError'
        '404':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseError'
        '500':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseError'
      security:
        - bearer: []
      summary: ''
      tags:
        - Users
  /users/members/stats:
    get:
      operationId: getUserMembersStats
      parameters:
        - name: fromDate
          required: false
          in: query
          schema:
            format: date
            type: string
        - name: endDate
          required: false
          in: query
          schema:
            format: date
            type: string
        - name: type
          required: false
          in: query
          schema:
            enum:
              - ACCULUMATED
              - PER_YEAR
            type: string
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserStat'
        '400':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseError'
        '403':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseError'
        '404':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseError'
        '500':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseError'
      security:
        - bearer: []
      summary: ''
      tags:
        - Users
  /deleted-roles:
    get:
      operationId: findDeletedRoles
      parameters:
        - name: role
          required: false
          in: query
          schema:
            enum:
              - ADMIN
              - REGION_MANAGER
              - COMMITTEE_MANAGER
              - ASSOCIATION_MANAGER
              - SIMPLE_USER
            type: string
        - name: page
          required: false
          in: query
          schema:
            type: number
            minimum: 1
            default: 1
        - name: pageSize
          required: false
          in: query
          schema:
            type: number
            default: 10
            minimum: 1
            maximum: 500
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DeletedRole'
        '400':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseError'
        '403':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseError'
        '404':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseError'
        '500':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseError'
      security:
        - bearer: []
      summary: ''
      tags:
        - Users
...
info:
  title: Jfds Api
  description: Jfds
  version: 0.0.1
  contact: {}
...
servers:
  - url: http://localhost:3000
components:
  securitySchemes:
    bearer:
      scheme: bearer
      bearerFormat: JWT
      type: http
  schemas:
...
    ApiResponseError:
      type: object
      properties:
        statusCode:
          type: number
        message:
          type: string
        error:
          type: string
      required:
        - statusCode
        - message
        - error
    Responsability:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - createdAt
        - updatedAt
    Committee:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        code:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - code
        - createdAt
        - updatedAt
    Region:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        code:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - code
        - createdAt
        - updatedAt
    Association:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        code:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - code
        - createdAt
        - updatedAt
    Sacrament:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - createdAt
        - updatedAt
...
    Count:
      type: object
      properties:
        value:
          type: number
      required:
        - value
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
        username:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        birthDate:
          type: string
          format: date
        address:
          type: string
        gender:
          type: string
          enum:
            - MALE
            - FEMALE
        role:
          type: string
          enum:
            - ADMIN
            - REGION_MANAGER
            - COMMITTEE_MANAGER
            - ASSOCIATION_MANAGER
            - SIMPLE_USER
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        nic:
          type: string
        photo:
          type: string
        apv:
          type: string
        responsability:
          $ref: '#/components/schemas/Responsability'
        committee:
          $ref: '#/components/schemas/Committee'
        region:
          $ref: '#/components/schemas/Region'
        association:
          $ref: '#/components/schemas/Association'
        sacrament:
          $ref: '#/components/schemas/Sacrament'
      required:
        - id
        - email
        - username
        - firstName
        - lastName
        - birthDate
        - address
        - gender
        - role
        - createdAt
        - updatedAt
    UploadeSuccessResponse:
      type: object
      properties:
        message:
          type: string
        fileName:
          type: string
      required:
        - message
        - fileName
    UpdateUser:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
        username:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        birthDate:
          type: string
          format: date
        address:
          type: string
        gender:
          type: string
          enum:
            - MALE
            - FEMALE
        role:
          type: string
          enum:
            - ADMIN
            - REGION_MANAGER
            - COMMITTEE_MANAGER
            - ASSOCIATION_MANAGER
            - SIMPLE_USER
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        nic:
          type: string
        apv:
          type: string
        responsabilityId:
          type: string
          format: uuid
        regionId:
          type: string
          format: uuid
        committeeId:
          type: string
          format: uuid
        associationId:
          type: string
          format: uuid
        sacramentId:
          type: string
          format: uuid
      required:
        - id
        - email
        - username
        - firstName
        - lastName
        - birthDate
        - address
        - gender
        - role
        - createdAt
        - updatedAt
    CreateUser:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
        username:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        birthDate:
          type: string
          format: date
        address:
          type: string
        gender:
          type: string
          enum:
            - MALE
            - FEMALE
        role:
          type: string
          enum:
            - ADMIN
            - REGION_MANAGER
            - COMMITTEE_MANAGER
            - ASSOCIATION_MANAGER
            - SIMPLE_USER
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        nic:
          type: string
        apv:
          type: string
        responsabilityId:
          type: string
          format: uuid
        regionId:
          type: string
          format: uuid
        committeeId:
          type: string
          format: uuid
        associationId:
          type: string
          format: uuid
        sacramentId:
          type: string
          format: uuid
        password:
          type: string
      required:
        - id
        - email
        - username
        - firstName
        - lastName
        - birthDate
        - address
        - gender
        - role
        - createdAt
        - updatedAt
        - password
    UserStat:
      type: object
      properties:
        year:
          type: number
        maleCount:
          type: number
        femaleCount:
          type: number
        totalCount:
          type: number
      required:
        - year
        - maleCount
        - femaleCount
        - totalCount
    DeletedRole:
      type: object
      properties:
        id:
          type: string
        role:
          type: string
          enum:
            - ADMIN
            - REGION_MANAGER
            - COMMITTEE_MANAGER
            - ASSOCIATION_MANAGER
            - SIMPLE_USER
        createdAt:
          type: string
          format: date-time
        user:
          $ref: '#/components/schemas/User'
      required:
        - id
        - role
        - createdAt
        - user
...